add import for json decks

+++ api/services/learning/Learning.Application/DTOs/Decks/DeckDto.cs
@@ -5,4 +5,4 @@ public record DeckDto(
         string UserEmail,
         string Topic,
         bool IsStrict,
-        int WordCount);
+        long WordCount);
